# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Traffic Peak typing platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tests:
    get:
      operationId: listTests
      tags: [tests]
      summary: Get recent test results for current user
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: mode
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of test results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TestResult"
    post:
      operationId: submitTest
      tags: [tests]
      summary: Submit a completed typing test
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestResultInput"
      responses:
        "201":
          description: Test result saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResult"
  /tests/{id}:
    get:
      operationId: getTest
      tags: [tests]
      summary: Get a single test result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestResult"
        "404":
          description: Not found

  # ── Stats / Dashboard ─────────────────────────────────────────────────
...
  /users/{username}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get a public user profile
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
  /users/me/profile:
    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── HR Assessments ────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, username, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        name:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string

    # ── Test Results ──────────────────────────────────────────────────
    TestResultInput:
      type: object
      required: [wpm, cpm, accuracy, duration, mode, errorCount, charCount]
      properties:
        wpm:
          type: number
        cpm:
          type: number
        accuracy:
          type: number
        duration:
          type: integer
        mode:
          type: string
        errorCount:
          type: integer
        charCount:
          type: integer
        language:
          type: string
    TestResult:
      type: object
      required: [id, wpm, cpm, accuracy, duration, mode, errorCount, charCount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        wpm:
          type: number
        cpm:
          type: number
        accuracy:
          type: number
        duration:
          type: integer
        mode:
          type: string
        errorCount:
          type: integer
        charCount:
          type: integer
        language:
          type: ["string", "null"]
        createdAt:
          type: string

    # ── Stats ─────────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      required: [username, bestWpm, avgWpm, avgAccuracy, totalTests, createdAt]
      properties:
        username:
          type: string
        name:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bestWpm:
          type: number
        avgWpm:
          type: number
        avgAccuracy:
          type: number
        totalTests:
          type: integer
        createdAt:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        avatarUrl:
          type: string

    # ── HR Assessments ─────────────────────────────────────────────────
...